// src/client/api/models/CompanyEventCreate.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
import type { CompanyEventEnum } from './CompanyEventEnum';
export type CompanyEventCreate = {
    title: string;
    date: string;
    event_type: CompanyEventEnum;
};