// src/client/api/models/CompanyEventUpdate.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
import type { CompanyEventEnum } from './CompanyEventEnum';
export type CompanyEventUpdate = {
    title: string;
    date: string;
    event_type: CompanyEventEnum;
};